Preserve tool downloader abstractions

Copilot-Session: 164ae8b0-7bdd-40e6-aa03-69ee79fc0488
