+++ src/Cli/dotnet/NugetPackageDownloader/INuGetPackageDownloader.cs
@@ -43,4 +43,10 @@ internal interface INuGetPackageDownloader
     Task<(NuGetVersion version, PackageSource source)> GetBestPackageVersionAndSourceAsync(PackageId packageId,
         VersionRange versionRange,
         PackageSourceLocation packageSourceLocation = null);
+
+    Task<(NuGetVersion version, PackageSource source)> GetBestPackageVersionAndSourceAsync(
+        PackageId packageId,
+        VersionRange versionRange,
+        PackageSourceLocation packageSourceLocation,
+        CancellationToken cancellationToken);
 }
