+++ test/Microsoft.DotNet.Tools.Tests.ComponentMocks/ToolPackageDownloaderMock.cs
@@ -335,6 +335,30 @@ namespace Microsoft.DotNet.Tools.Tests.ComponentMocks
             return (NuGetVersion.Parse(feedPackage.Version), new PackageSource("http://mock-feed", "MockFeed"));
         }
 
+        public Task<(NuGetVersion version, PackageSource source)> GetNuGetVersionAsync(
+            PackageLocation packageLocation,
+            PackageId packageId,
+            VerbosityOptions verbosity,
+            VersionRange versionRange,
+            RestoreActionConfig restoreActionConfig,
+            CancellationToken cancellationToken)
+        {
+            cancellationToken.ThrowIfCancellationRequested();
+            return Task.FromResult(GetNuGetVersion(
+                packageLocation,
+                packageId,
+                verbosity,
+                versionRange,
+                restoreActionConfig));
+        }
+
+        public bool TryGetBestDownloadedTool(
+            PackageId packageId,
+            VersionRange versionRange,
+            string? targetFramework,
+            VerbosityOptions verbosity,
+            [NotNullWhen(true)] out IToolPackage? toolPackage) => throw new NotImplementedException();
+
         public bool TryGetDownloadedTool(PackageId packageId, NuGetVersion packageVersion, string? targetFramework, VerbosityOptions verbosity, [NotNullWhen(true)] out IToolPackage? toolPackage) => throw new NotImplementedException();
 
         private class TestToolPackage : IToolPackage
