+++ test/dotnet-aot.Tests/WorkloadUpdateTests.cs
@@ -283,5 +283,11 @@ public class WorkloadUpdateTests
             PackageId packageId,
             VersionRange versionRange,
             PackageSourceLocation? packageSourceLocation = null) => throw new NotSupportedException();
+
+        public Task<(NuGetVersion version, PackageSource source)> GetBestPackageVersionAndSourceAsync(
+            PackageId packageId,
+            VersionRange versionRange,
+            PackageSourceLocation? packageSourceLocation,
+            CancellationToken cancellationToken) => throw new NotSupportedException();
     }
 }
