+++ test/dotnet.Tests/CommandTests/Workload/Install/FailingNuGetPackageInstaller.cs
@@ -46,6 +46,12 @@ namespace Microsoft.DotNet.Cli.NuGetPackageDownloader
         public Task<(NuGetVersion version, PackageSource source)> GetBestPackageVersionAndSourceAsync(PackageId packageId,
             VersionRange versionRange, PackageSourceLocation packageSourceLocation = null) => throw new NotImplementedException();
 
+        public Task<(NuGetVersion version, PackageSource source)> GetBestPackageVersionAndSourceAsync(
+            PackageId packageId,
+            VersionRange versionRange,
+            PackageSourceLocation packageSourceLocation,
+            CancellationToken cancellationToken) => throw new NotImplementedException();
+
         public Task<string> GetPackageUrl(PackageId packageId,
             NuGetVersion packageVersion,
             PackageSourceLocation packageSourceLocation = null,
